Delete the DX9 vertex decoder, use the "GL" one instead (now common)
