+++ GPU/GLES/VertexShaderGenerator.h
@@ -52,6 +52,3 @@ bool CanUseHardwareTransform(int prim);
 
 void ComputeVertexShaderID(VertexShaderID *id, u32 vertexType, int prim, bool useHWTransform);
 void GenerateVertexShader(int prim, u32 vertexType, char *buffer, bool useHWTransform);
-
-// Collapse to less skinning shaders to reduce shader switching, which is expensive.
-int TranslateNumBones(int bones);
